LPS-165169 Inline and remove unused classes
